mongo.py
Problem:
We store the sales data of an e-commerce store in MongoDB and predict future sales of the product.

Project Contents:
We store the data in MongoDB.
We retrieve the data from MongoDB with Python.
We process and clean the data with Pandas.
Linear Regression or Time Series Analysis.
We store and/or visualize the results in MongoDB.